requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==6.0.1
django-allauth==65.14.0
gunicorn==25.1.0
idna==3.11
packaging==26.0
pillow==12.1.0
pycparser==3.0
PyJWT==2.10.1
python-dotenv==1.2.1
razorpay==2.0.0
requests==2.32.5
sqlparse==0.5.5
stripe==14.3.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
